weakly check in - not working
